fix: bugs de incompatibilidade
