+++ src/components/LoginCard/TermsOfUse/index.tsx
@@ -1,4 +1,5 @@
-import React, { useEffect, useRef } from 'react';
+import React from 'react';
+import CloseModalEsc from '../CloseModalEsc/index';
 import {
     ModalContent,
     ModalTitle,
@@ -14,34 +15,7 @@ interface TermsOfUseProps {
 }
 
 const TermsOfUse: React.FC<TermsOfUseProps> = ({ isOpen, onClose }) => {
-    const modalRef = useRef<HTMLDivElement>(null);
-
-    useEffect(() => {
-        const handleEsc = (event: KeyboardEvent) => {
-            if (event.key === 'Escape') {
-                onClose();
-            }
-        };
-
-        const handleClickOutside = (event: MouseEvent) => {
-            if (
-                modalRef.current &&
-                !modalRef.current.contains(event.target as Node)
-            ) {
-                onClose();
-            }
-        };
-
-        if (isOpen) {
-            document.addEventListener('keydown', handleEsc);
-            document.addEventListener('mousedown', handleClickOutside);
-        }
-
-        return () => {
-            document.removeEventListener('keydown', handleEsc);
-            document.removeEventListener('mousedown', handleClickOutside);
-        };
-    }, [isOpen, onClose]);
+    const modalRef = CloseModalEsc(isOpen, onClose);
 
     if (!isOpen) return null;
